Alterando configurações de autenticação.

+++ database/migrations/2025_06_21_185159_create_tokens_table.php
@@ -11,14 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('tokens', function (Blueprint $table) {
-            $table->uuid('id')->primary();
-            $table->char('token', 36);
-            $table->string('email', 100);
-            $table->date('data_criacao');
-            $table->date('data_expiracao');
-            $table->foreign('email')->references('email')->on('users')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('tokens')) {
+            Schema::create('tokens', function (Blueprint $table) {
+                $table->uuid('id')->primary();
+                $table->char('token', 36);
+                $table->string('email', 100);
+                $table->date('data_criacao');
+                $table->date('data_expiracao');
+                $table->foreign('email')->references('email')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
